fix(pipeline): move monitoring loop in frontend
State of the pipeline webservice is checked on the frontend
side every second with a setInterval in the store
Additionnaly, some utility functions are moved from
pipeline2 class file to a utils file.
