+++ src/renderer/store/index.tsx
@@ -70,6 +70,32 @@ export function WindowStoreProvider({ children }) {
     }, [])
 
     useEffect(() => {
+        const interval = setInterval(async () => {
+            if (pipeline.runningWebservice) {
+                fetch(`${baseurl(pipeline.runningWebservice)}/alive`)
+                    .then((value: Response) => {
+                        if (pipeline.status != PipelineStatus.RUNNING) {
+                            pipeline.status = PipelineStatus.RUNNING
+                            setPipelineState({
+                                ...pipeline,
+                                status: PipelineStatus.RUNNING,
+                            })
+                        }
+                    })
+                    .catch((reason) => {
+                        // Change the status to error only if the previous one was the running status
+                        // (because the previous could be the starting one, and in this cas i don't want changes
+                        // as it may be only the delay of booting up returning an error)
+                        if (pipeline.status == PipelineStatus.RUNNING) {
+                            pipeline.status = PipelineStatus.STOPPED
+                            setPipelineState({
+                                ...pipeline,
+                                status: PipelineStatus.STOPPED,
+                            })
+                        }
+                    })
+            }
+        }, 1000)
         if (pipeline.status == PipelineStatus.RUNNING && scripts.length == 0) {
             console.log('getting scripts (useEffect)')
             getScripts(
